Move Xamarin problematic utils classes to utils.storage
Spent all afternoon trying to exclude failing bindings by class/interface names instead of package with no luck.
