+++ sdk/analytics/src/main/java/com/microsoft/sonoma/analytics/channel/SessionTracker.java
@@ -8,8 +8,8 @@ import com.microsoft.sonoma.analytics.ingestion.models.StartSessionLog;
 import com.microsoft.sonoma.core.channel.Channel;
 import com.microsoft.sonoma.core.ingestion.models.Log;
 import com.microsoft.sonoma.core.utils.SonomaLog;
-import com.microsoft.sonoma.core.utils.StorageHelper;
 import com.microsoft.sonoma.core.utils.UUIDUtils;
+import com.microsoft.sonoma.core.utils.storage.StorageHelper;
 
 import java.util.HashSet;
 import java.util.Map;
