+++ sdk/core/src/androidTest/java/com/microsoft/sonoma/core/utils/storage/SQLiteUtilsAndroidTest.java
@@ -1,4 +1,4 @@
-package com.microsoft.sonoma.core.utils;
+package com.microsoft.sonoma.core.utils.storage;
 
 import org.junit.Test;
