+++ sdk/core/src/main/java/com/microsoft/sonoma/core/AbstractSonomaFeature.java
@@ -8,7 +8,7 @@ import android.support.annotation.NonNull;
 import com.microsoft.sonoma.core.channel.Channel;
 import com.microsoft.sonoma.core.ingestion.models.json.LogFactory;
 import com.microsoft.sonoma.core.utils.SonomaLog;
-import com.microsoft.sonoma.core.utils.StorageHelper;
+import com.microsoft.sonoma.core.utils.storage.StorageHelper;
 
 import java.util.Map;
