+++ sdk/crashes/src/test/java/com/microsoft/sonoma/crashes/CrashesTest.java
@@ -12,8 +12,8 @@ import com.microsoft.sonoma.core.ingestion.models.json.LogFactory;
 import com.microsoft.sonoma.core.ingestion.models.json.LogSerializer;
 import com.microsoft.sonoma.core.utils.PrefStorageConstants;
 import com.microsoft.sonoma.core.utils.SonomaLog;
-import com.microsoft.sonoma.core.utils.StorageHelper;
 import com.microsoft.sonoma.core.utils.UUIDUtils;
+import com.microsoft.sonoma.core.utils.storage.StorageHelper;
 import com.microsoft.sonoma.crashes.ingestion.models.ManagedErrorLog;
 import com.microsoft.sonoma.crashes.ingestion.models.StackFrame;
 import com.microsoft.sonoma.crashes.ingestion.models.json.ManagedErrorLogFactory;
