+++ versions.gradle
@@ -1,8 +1,8 @@
 // Version constants
 
 ext {
-    versionCode = 3
-    versionName = '0.1.2'
+    versionCode = 4
+    versionName = '0.1.3'
     minSdkVersion = 15
     targetSdkVersion = 25
     compileSdkVersion = 25
